requirements.txt
Django>=2.2.2
django-ckeditor>=5.7.1
Pillow>=6.0.0